+++ ykdl/extractors/pptv.py
@@ -17,6 +17,7 @@ from xml.dom.minidom import parseString
 
 def lshift(a, b):
     return (a << b) & 0xffffffff
+
 def rshift(a, b):
     if a >= 0:
         return a >> b
@@ -37,6 +38,7 @@ def le32_pack(b_str):
         result |= (ord(b_str[2]) << 16)
         result |= (ord(b_str[3]) << 24)
         return result
+
 def tea_core(data, key_seg):
     delta = 2654435769
 
@@ -122,7 +124,9 @@ def parse_pptv_xml(dom):
         item_mlist.append(item_meta)
 
     dt_list = get_elem(dom, 'dt')
-    dragdata_list = get_elem(dom, 'dragdata')
+    # 部分视频没有dragdata标签
+    # dragdata_list = get_elem(dom, 'dragdata')
+    dragdata_list = get_elem(dom, 'dragdata') or get_elem(dom, 'drag')
 
     stream_mlist = []
     for dt in dt_list:
@@ -153,11 +157,12 @@ def merge_meta(item_mlist, stream_mlist, segs_mlist):
     streams = {}
     for i in range(len(segs_mlist)):
         streams[str(i)] = {}
-
     for item in item_mlist:
         stream = streams[item[0]]
         stream['rid'] = item[1]
-        stream['size'] = item[2]
+        # 无size情况,如无需求,指定默认值
+        # stream['size'] = item[2]
+        stream['size'] = item[2] or 12653713
         stream['res'] = item[3]
 
     for s in stream_mlist:
@@ -170,7 +175,6 @@ def merge_meta(item_mlist, stream_mlist, segs_mlist):
         stream = streams[seg[0]]
         stream['segs'] = seg[1]
         stream['segs_size'] = seg[2]
-
     return streams
 
 
@@ -183,7 +187,9 @@ def make_url(stream):
     src = []
     for i, seg in enumerate(stream['segs']):
         url = 'http://{}/{}/{}?key={}&k={}'.format(host, i, rid, key, key_expr)
-        url += '&fpp.ver=1.3.0.23&type=web.fpp'
+        # type修改成ppbox.launcher
+        # url += '&fpp.ver=1.3.0.23&type=web.fpp'
+        url += '&fpp.ver=1.3.0.23&type=ppbox.launcher'
         src.append(url)
     return src
 
@@ -194,14 +200,17 @@ class Pptv(VideoExtractor):
 
     def prepare(self):
         info = VideoInfo(self.name)
-        html = get_content(self.url)
-        self.vid = match1(html, 'webcfg\s*=\s*{"id":\s*(\d+)')
-        param = "type%3dweb.fpp%26ahl_ver%3d1%26ahl_random%3d6c2b3072426c42253c754c4460624b76%26ahl_signa%3d8544ec938b8b6e4153320931d5079e7aadfbed5855a5ccc40c66d470338b7056%26userType%3d0%26o%3d0"
-        xml = get_content('http://web-play.pptv.com/webplay3-0-{}.xml?version=4&param={}&type=web.fpp&appplt=flp&appid=pptv.flashplayer.vod&appver=3.4.2.32'.format(self.vid,param))
+        self.vid = match1(self.url, '\.pptv\.com/vod/(\d+)')
+        if not self.vid:
+            html = get_content(self.url)
+            self.vid = match1(html, 'webcfg\s*=\s*{"id":\s*(\d+)')
+         # API修改
+        # param = "type%3dppbox.launcher%26ahl_ver%3d1%26ahl_random%3d6c2b3072426c42253c754c4460624b76%26ahl_signa%3d8544ec938b8b6e4153320931d5079e7aadfbed5855a5ccc40c66d470338b7056%26userType%3d0%26o%3d0"
+        # xml = get_content('http://web-play.pptv.com/webplay3-0-{}.xml?version=4&param={}&type=web.fpp&appplt=flp&appid=pptv.flashplayer.vod&appver=3.4.2.32'.format(self.vid,param))
+        xml = get_content('https://web-play.pptv.com/webplay3-0-{}.xml?zone=8&version=4&username=&ppi=302c3333&type=ppbox.launcher&pageUrl=http%3A%2F%2Fv.pptv.com&o=0&referrer=&kk=&scver=1&appplt=flp&appid=pptv.flashplayer.vod&appver=3.4.3.3&nddp=1'.format(self.vid))
         dom = parseString(compact_bytes(xml, 'utf-8'))
         info.title, m_items, m_streams, m_segs = parse_pptv_xml(dom)
         xml_streams = merge_meta(m_items, m_streams, m_segs)
-
         for stream_id in xml_streams:
             stream_data = xml_streams[stream_id]
             src = make_url(stream_data)