+++ ykdl/util/html.py
@@ -7,6 +7,8 @@ from ykdl.compact import Request, urlopen
 
 from .match import match1
 
+default_proxy_handler = []
+
 fake_headers = {
     'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
     'Accept-Encoding': 'gzip, deflate',
@@ -69,6 +71,12 @@ def get_content(url, headers=fake_headers, data=None, charset = None):
     resheader = response.info()
     if 'Content-Encoding' in resheader:
         content_encoding = resheader['Content-Encoding']
+    elif hasattr(resheader, 'get_payload'):
+        payload = resheader.get_payload()
+        if isinstance(payload, str):
+            content_encoding =  match1(payload, r'Content-Encoding:\s*([\w-]+)')
+        else:
+            content_encoding = None
     else:
         content_encoding = None
     if content_encoding == 'gzip':
@@ -83,10 +91,10 @@ def get_content(url, headers=fake_headers, data=None, charset = None):
     if charset is None:
         if 'Content-Type' in resheader:
             charset = match1(resheader['Content-Type'], r'charset=([\w-]+)')
-        charset = charset or match1(str(data), r'charset=\"([^\"]+)', 'charset=([^"]+)') or 'utf-8'
+        charset = charset or match1(str(data), r'charset=\"([\w-]+)', 'charset=([\w-]+)') or 'utf-8'
     logger.debug("get_content> Charset: " + charset)
     try:
-        data = data.decode(charset)
+        data = data.decode(charset, errors='replace')
     except:
         logger.warning("wrong charset for {}".format(url))
     return data