+++ CHANGELOG.rst
@@ -1,6 +1,32 @@
 Change Log for ykdl
 ===================
 
+1.6.3
+-------
+
+- update Sohu, mgtv, sina, etc a lot by @SeaHOH
+
+
+1.6.2
+-------
+
+- update QQ, bilibili, panda a lot by @SeaHOH
+- misc updates
+
+1.6.1
+-------
+
+- fix youku/tudou, QQ, mgtv, 163, bilibili by @SeaHOH
+
+1.6.0
+-------
+
+- new internal proxy for slow video playback. by @SeaHOH
+- update youku, many bug fixed.
+- update bilibili, new api, playlist bug fixed.
+- update HTTP redirection for t.cn
+- use cryptodome as default.
+
 1.5.5
 -------